requirements.txt
apache-airflow>=1.10.2
sentry-sdk>=0.5.4
celery>=4.3.0
